Fix: Add time buffer and dubug option in travis build
After being started, `Couchdb` can not bind port 5984 by the time its `userName` and `password` is being set. So the `curl` command fails with error code 7.

`sudo updatedb` is a buffer command in `travis` `install` phase to give `couchdb` time to bind 5984 port. For this `sudo` is made required at top.

`sudo netstat -tuplen` is added to list all the active ports for debugging purpose in case of `curl` failure.